requirements.txt
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
comtypes==1.4.13
google-auth==2.43.0
google-genai==1.52.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.12.4
pydantic_core==2.41.5
pypiwin32==223
python-dotenv==1.2.1
pyttsx3==2.99
pywin32==311
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
